internal/task: ignore okay-after-beta1 when releasing beta 2+
The original intention for the okay-after-beta1 label
was to apply up to and including the release of beta 1.
It did not have any effect on beta 2, beta 3, etc., in releasebot.

This is because we can't see that far into the future when making
decisions of whether to apply that label to some release-blockers.
It exists to let us prioritize what's needed to unblock the first
pre-release of the upcoming major Go release.

This change makes it so again.

Change-Id: I3304cba329b1044dffb7c8c17ad9c92d690cbb55
Reviewed-on: https://go-review.googlesource.com/c/build/+/453981
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>