cmd/gomote: push local VERSION file if it exists
When the gomote push command is issued, a VERSION file is
generated and pushed to the gomote instance. This change
checks if there is a local VERSION file and uses the local
VERSION file if it exists.

Fixes golang/go#57068

Change-Id: I924b2b946d371c4204460b8e9d1cad0a3a6fddd6
Reviewed-on: https://go-review.googlesource.com/c/build/+/460775
Run-TryBot: Carlos Amedee <carlos@golang.org>
Reviewed-by: Heschi Kreinick <heschi@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>