cmd/relui: add await step for goimports CL, remove ad-hoc workflow
The ad-hoc workflow was successfully used to generate CL 464875 and
won't be needed anymore, so drop it.

Add a "await CL" step so there's a lightweight indicator of when the
goimports CL is submitted, like we have for the DL CL and version CL.
It's a leaf task that doesn't block any other Go release process work,
but helps make it so that when the overall workflow is completed, we
know there's nothing more that needs attention.

Fixes golang/go#54377.

Change-Id: I00cfd851677e0c6c0041f4f2b913e7740b3f7ab4
Reviewed-on: https://go-review.googlesource.com/c/build/+/464915
Reviewed-by: Carlos Amedee <carlos@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>